chore: rename gits_noise.inl
