+++ src/runtime/gits_noise.h
